requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymodbus==3.9.2
paho-mqtt==2.1.0
pandas==2.2.2
python-dateutil==2.9.0.post0
requests==2.31.0
-e ../timeauthority-pkg
